tools/requirements.txt
pygame>=2.6,<3
pyserial>=3.5,<4
bleak>=0.20.0  # For BLE support